refactoring: move title-positon maps into own file
